+++ src/component/table/table.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react'
 import { BookTable } from './table'
 import { Book } from '../../bookshelf/book/book'
 import { algorithms, books } from '../../support/book-fixtures'
-import { fn } from '@storybook/test'
+import { action } from '@storybook/addon-actions'
 
 const meta = {
     title: 'Table',
@@ -12,7 +12,7 @@ export default meta
 type Story = StoryObj<typeof BookTable>
 
 function renderFunction(books: Array<Book>) {
-    return () => <BookTable books={books} onBookClick={fn()} />
+    return () => <BookTable books={books} onBookClick={action('onBookClick')} />
 }
 
 export const Primary: Story = {