+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 For a full diff see [`0.11.0...main`][0.11.0...main].
 
+### Added
+
+- Open book note directly from gallery/table view ([#8])\
+  You can now open a book's note directly by holding a modifier key while clicking a book in the gallery or table view
+  (CMD on macOS, CTRL on Windows and Linux). Thanks to [@soundslikeinfo] for the suggestion!
+
 ## [`0.11.0`][0.11.0] - 2025-04-15
 
 For a full diff see [`0.10.0...0.11.0`][0.10.0...0.11.0].
@@ -175,3 +181,5 @@ For a full diff see [`0.2.2...0.3.0`][0.2.2...0.3.0].
 [0.9.0...0.10.0]: https://github.com/weph/obsidian-bookshelf/compare/0.9.0...0.10.0
 [0.10.0...0.11.0]: https://github.com/weph/obsidian-bookshelf/compare/0.10.0...0.11.0
 [0.11.0...main]: https://github.com/weph/obsidian-bookshelf/compare/0.11.0...main
+[#8]: https://github.com/weph/obsidian-bookshelf/issues/8
+[@soundslikeinfo]: https://github.com/soundslikeinfo